requirements.txt
Flask==2.0.2
python-dotenv==0.19.2
openai==0.28.0
